+++ src/main/java/com/adl/et/telco/mvno/productcatalog/application/controller/CategoryController.java
@@ -8,10 +8,12 @@ import com.adl.et.telco.mvno.productcatalog.application.config.Constants;
 import com.adl.et.telco.mvno.productcatalog.application.transport.request.entities.CategoryCreateRequestEntity;
 import com.adl.et.telco.mvno.productcatalog.application.transport.request.entities.CategoryUpdateRequestEntity;
 import com.adl.et.telco.mvno.productcatalog.domain.entities.Category;
+import io.swagger.annotations.Api;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@Api(tags = "category")
 @RequestMapping("${app.context.absolute}/" + Constants.UrlConstants.CATEGORY_RESOURCE)
 public class CategoryController extends BaseResourceController<Category,
         CategoryCreateRequestEntity,